requirements.txt
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
ez-setup==0.9
idna==2.10
pdfminer==20191125
pdfminer.six==20201018
pycparser==2.20
pycryptodome==3.9.8
python-dotenv==0.15.0
requests==2.24.0
six==1.15.0
sortedcontainers==2.2.2
urllib3==1.25.11
